requirements.txt
accelerate==0.33.0
aiofiles==23.2.1
aiohappyeyeballs==2.3.4
aiohttp==3.10.0
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
anyio==3.7.1
appdirs==1.4.4
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
bitsandbytes==0.43.3
black==24.8.0
blinker==1.8.2
blobfile==2.1.1
boltons==21.0.0
bracex==2.5
Brotli==1.1.0
cachetools==5.4.0
certifi @ file:///croot/certifi_1720453481653/work/certifi
chardet==5.2.0
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.1.7
click-option-group==0.5.6
codeshield==1.0.1
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.2.1
cycler==0.12.1
dataclasses-json==0.6.7
datasets==2.20.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.8
distro==1.9.0
exceptiongroup==1.2.2
executing==2.0.1
face==22.0.0
fairscale==0.4.13
fastapi==0.112.0
ffmpy==0.4.0
filelock @ file:///croot/filelock_1700591183607/work
fire==0.6.0
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.5.0
gitdb==4.0.11
GitPython==3.1.43
glom==22.1.0
gmpy2 @ file:///tmp/build/80754af9/gmpy2_1645438755360/work
googleapis-common-protos==1.63.2
gradio==3.50.2
gradio_client==0.6.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.5
humanfriendly==10.0
idna @ file:///croot/idna_1714398848350/work
importlib_metadata==7.1.0
importlib_resources==6.4.0
inflate64==1.0.0
ipython==8.18.1
jedi==0.19.1
Jinja2 @ file:///croot/jinja2_1716993405101/work
joblib==1.4.2
json_strong_typing==0.3.4
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
langchain==0.2.12
langchain-community==0.2.11
langchain-core==0.2.28
langchain-text-splitters==0.2.2
langsmith==0.1.96
llama-recipes==0.0.3
loralib==0.1.2
lxml==4.9.4
markdown-it-py==3.0.0
MarkupSafe @ file:///croot/markupsafe_1704205993651/work
marshmallow==3.21.3
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
mpmath @ file:///croot/mpmath_1690848262763/work
multidict==6.0.5
multiprocess==0.70.16
multivolumefile==0.2.3
mypy-extensions==1.0.0
networkx @ file:///croot/networkx_1717597493534/work
numpy @ file:///croot/numpy_and_numpy_base_1708638617955/work/dist/numpy-1.26.4-cp39-cp39-linux_x86_64.whl#sha256=6094eeedd869502faa0fd0a8c5ad3a70c5779be06ddd1feb7627e5c212fac420
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu12==12.1.105
ollama==0.3.1
openai==1.2.0
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-http==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-instrumentation-requests==0.46b0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
opentelemetry-util-http==0.46b0
optimum==1.17.1
orjson==3.10.6
packaging==24.1
pandas==2.1.4
parso==0.8.4
pathspec==0.12.1
peewee==3.17.6
peft==0.12.0
pexpect==4.9.0
pillow @ file:///croot/pillow_1721059439630/work
platformdirs==4.2.2
prompt_toolkit==3.0.47
protobuf==4.25.4
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py7zr==0.21.1
pyarrow==17.0.0
pyarrow-hotfix==0.6
pybcj==1.0.2
pycryptodomex==3.20.0
pydantic==1.10.13
pydantic_core==2.18.2
pydeck==0.9.1
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.1.2
pyppmd==1.1.0
PySocks @ file:///tmp/build/80754af9/pysocks_1605305812635/work
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML @ file:///croot/pyyaml_1698096049011/work
pyzstd==0.16.1
referencing==0.35.1
regex==2024.7.24
requests @ file:///croot/requests_1721410876868/work
rich==13.7.1
rpds-py==0.19.1
ruamel.yaml==0.17.40
ruamel.yaml.clib==0.2.8
ruff==0.5.6
safetensors==0.4.3
scikit-learn==0.24.2
scipy==1.5.4
semantic-version==2.10.0
semgrep==1.83.0
sentencepiece==0.2.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.32
stack-data==0.6.3
starlette==0.37.2
streamlit==1.37.0
sympy @ file:///croot/sympy_1701397643339/work
tabulate==0.9.0
tenacity==8.5.0
termcolor==2.4.0
texttable==1.7.0
threadpoolctl==3.5.0
tiktoken==0.4.0
tokenize-rt==6.0.0
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.1
torch==2.4.0
torchaudio==2.4.0
torchvision==0.19.0
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.43.4
triton==3.0.0
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2024.1
urllib3 @ file:///croot/urllib3_1718912636303/work
uvicorn==0.30.5
watchdog==4.0.1
wcmatch==8.5.2
wcwidth==0.2.13
websockets==11.0.3
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.19.2
